requirements.txt
streamlit
pandas
scikit-learn
sentence-transformers
fuzzywuzzy[speedup]
requests
python-dotenv
joblib